requirements.txt
memray
streamlit-redis==0.0.3
streamlit-1.28.1-py2.py3-none-any.whl
